+++ src/components/Header/Header.tsx
@@ -3,8 +3,10 @@ import { Link } from 'react-router-dom'
 import logo from '../../images/logo.png';
 import { useAuth } from '../../hooks/useAuth';
 import { useAppDispatch } from '../../hooks/reduxHooks';
-import { logoutAction } from '../../redux/actions/authorizationActions';
-export default function Header() {
+import { signOut } from 'firebase/auth';
+import { auth } from '../../firebase';
+import { removeUser } from '../../redux/store/slices/userSlice';
+export function Header() {
   const { isAuth, email } = useAuth();
   const dispatch = useAppDispatch();
   return (
@@ -25,7 +27,8 @@ export default function Header() {
             <button
               className={styles.navItem}
               onClick={() => {
-                dispatch(logoutAction())
+                signOut(auth)
+                dispatch(removeUser())
               }}>
               Выйти
             </button>
