+++ src/pages/LoginPage/LoginPage.tsx
@@ -1,17 +1,25 @@
+import { useNavigate } from 'react-router-dom';
+import { signInWithEmailAndPassword } from "firebase/auth";
 import { Form } from '../../components/Form';
-import { useAppDispatch } from '../../hooks/reduxHooks';
-import { loginAction } from '../../redux/actions/authorizationActions';
+import { auth } from '../../firebase';
+
 
 export default function LoginPage() {
-  const dispatch = useAppDispatch()
+  const navigate = useNavigate(); 
   const handleLogin = (email: string, password: string) => {
-    dispatch(loginAction({email, password}))
+    signInWithEmailAndPassword(auth, email, password)
+      .then(({user}) => {
+        navigate('/');
+      })
+      .catch(() => alert('Неверно введены данные!'))
+
+
   }
   return (
     <Form
       isLogin={true}
       title='Войти'
-      handleClick={handleLogin}
+      handleClick={handleLogin} 
     />
   );
 }
