+++ src/utils/isEmailValid.ts
@@ -1,4 +1,4 @@
 const EMAIL_REGEXP = /^(([^<>()[\].,;:\s@"]+(\.[^<>()[\].,;:\s@"]+)*)|(".+"))@(([^<>()[\].,;:\s@"]+\.)+[^<>()[\].,;:\s@"]{2,})$/iu;
-export function isEmailValid(value: string): boolean {
+export function isEmailValid(value: string) {
   return EMAIL_REGEXP.test(value);
 }
